Add init script to remove dev toolbar:
When Playwright reuses a dev server that was already spawned before the
tests start, there is no way to set the env var required to disabled the
Astro dev toolbar. Thus, an init script is required to remove it, and
watch any changes on the page to ensure it doesn't come back.